# Avoid `as_tibble()` in `group_data.tbl_df()`